Expose the target ContainerConfiguration to Core
Add a new lifecycle phase 'init' for the Chameleon container which
is called before BeforeSetup. Remove the Chameleon container
configuration properties from the source configuration
and populate the Chameleon configuration object. During init,
resolve the target adapter and it's dependencies and
download the target distribution if required.

When BeforeSetup is called, Chameleon return the target's
ContainerConfiguration and a fullly populated target is handed back
in setup(). From there we move on as normal.

This should allow 2 things:

* Extensions can reason about the target container configuration
as normal with out Chameleon getting in the way

* ContainerController.start(Config) will operate on the target
configuration, not the Chameleon configuration.

fixes #12