adds the possibility to use maven from a local installation
this will basibally tell shrinkwrap resolver not to use a specific
maven installation, preventing it to download it from the internet
and, instead, using it from the local (either from maven.home or
M2_HOME env variables).

This behavior is specified here:
https://maven.apache.org/shared/maven-invoker/apidocs/org/apache/maven/shared/invoker/Invoker.html#setMavenHome-java.io.File-

This feature is useful when running arquillian tests from within
an environment with no internet access.